Weather Forecast Android App
The App that tells the weather forecast in both degree Celsius and
Farenheit. It also is integrated with facebook sharing wherein you can
share the weather forecast on facebook. Also integrated with the AERIES
maps it also helps you provide the map view of the weather. It provides
us with the data of next 48hours and next 7 days.